# Faster pointwise fusion graph pass