--- a/src/imperative/simple_partition_pass.cc
+++ b/src/imperative/simple_partition_pass.cc
@@ -33,92 +33,184 @@ namespace exec {
 
 namespace detail {
 
-void MergeSets(std::vector<Interval>** my_set,
-               std::vector<Interval>* other_set,
-               std::vector<std::unique_ptr<std::vector<Interval>>>* storage) {
+void PrintSets(const IntervalVec* const sets_p) {
+  if (sets_p == nullptr || sets_p->size() == 0) {
+    std::cout << "{}" << std::endl;
+    return;
+  }
+  const auto& sets = *sets_p;
+  std::cout << "{";
+  for (size_t i = 0; i < sets.size() - 1; ++i) {
+    std::cout << "[" << sets[i].first << "," << sets[i].second << "], ";
+  }
+  std::cout << "[" << sets[sets.size()-1].first << "," << sets[sets.size()-1].second << "]}" << std::endl;
+}
+
+const IntervalVec* LargerSet(const IntervalVec* const first,
+                             const IntervalVec* const second) noexcept {
+  const IntervalVec* ret = nullptr;
+  auto first_iter = first->begin();
+  auto second_iter = second->begin();
+  while (first_iter != first->end() &&
+         second_iter != second->end()) {
+    if (*first_iter == *second_iter) {
+      ++first_iter;
+      ++second_iter;
+    } else {
+      // Entry in first set not seen in the second set
+      if (first_iter->second < second_iter->first) {
+        if (ret == first || ret == nullptr) {
+          ret = first;
+          ++first_iter;
+        } else {
+          return nullptr;
+        }
+        continue;
+      }
+      // Entry in second set not seen in the first set
+      if (second_iter->second < first_iter->first) {
+        if (ret == second || ret == nullptr) {
+          ret = second;
+          ++second_iter;
+        } else {
+          return nullptr;
+        }
+        continue;
+      }
+      // Entry in first set fully encloses the entry in the second set
+      if (first_iter->first <= second_iter->first &&
+          first_iter->second >= second_iter->second) {
+        if (ret == first || ret == nullptr) {
+          ret = first;
+          ++second_iter;
+        } else {
+          return nullptr;
+        }
+        continue;
+      }
+      // Entry in second set fully encloses the entry in the first set
+      if (second_iter->first <= first_iter->first &&
+          second_iter->second >= first_iter->second) {
+        if (ret == second || ret == nullptr) {
+          ret = second;
+          ++first_iter;
+        } else {
+          return nullptr;
+        }
+        continue;
+      }
+      // Entries intersect but one is not fully enclosed in the other
+      return nullptr;
+    }
+  }
+  if (ret == nullptr) {
+    // The common part is the same
+    return second_iter == second->end() ? first : second;
+  } else {
+    if ((ret == first && second_iter == second->end()) ||
+        (ret == second && first_iter == first->end())) {
+      return ret;
+    }
+  }
+  return nullptr;
+}
+
+void MergeSets(const IntervalVec** const my_set,
+               const IntervalVec* const other_set,
+               std::vector<std::unique_ptr<const IntervalVec>>* const storage) noexcept {
   if ((*my_set == nullptr) || (*my_set)->size() == 0) {
     *my_set = other_set;
     return;
   }
-  if (other_set == nullptr || other_set->size() == 0 || other_set == (*my_set)) {
+  if (other_set == nullptr || other_set->size() == 0) {
     return;
   }
-  size_t current_interval = 0, current_other_interval = 0;
-  storage->emplace_back(std::make_unique<std::vector<Interval>>());
-  auto& new_set = storage->back();
+  auto* larger_set = LargerSet(*my_set, other_set);
+  if (larger_set != nullptr) {
+    *my_set = larger_set;
+    return;
+  }
+  auto my_iter = (*my_set)->cbegin();
+  auto other_iter = other_set->cbegin();
+  auto new_set = IntervalVec();
   int last_end = -10;  // less than -1
-  while (current_interval < (*my_set)->size() &&
-         current_other_interval < other_set->size()) {
-    const auto& mine = (**my_set)[current_interval];
-    const auto& other = (*other_set)[current_other_interval];
+  while (my_iter != (*my_set)->cend() &&
+         other_iter != other_set->cend()) {
+    const auto& mine = *my_iter;
+    const auto& other = *other_iter;
     if (other.second < mine.first - 1) {
       // other interval is before ours
       if (last_end >= other.first - 1) {
-        new_set->back().second = other.second;
+        new_set.back().second = other.second;
       } else {
-        new_set->emplace_back(other);
+        new_set.emplace_back(other);
       }
       last_end = other.second;
-      ++current_other_interval;
+      ++other_iter;
     } else if (other.first > mine.second + 1) {
       // other interval is after ours
       if (last_end >= mine.first - 1) {
-        new_set->back().second = std::max(mine.second, last_end);
+        new_set.back().second = std::max(mine.second, last_end);
       } else {
-        new_set->emplace_back(mine);
+        new_set.emplace_back(mine);
       }
-      last_end = new_set->back().second;
-      ++current_interval;
+      last_end = new_set.back().second;
+      ++my_iter;
     } else {
       // Intervals can be merged together
       Interval n(std::min(mine.first, other.first),
                  std::max(mine.second, other.second));
       if (last_end >= n.first - 1) {
-        new_set->back().second = n.second;
+        new_set.back().second = n.second;
       } else {
-        new_set->emplace_back(n);
+        new_set.emplace_back(n);
       }
       last_end = n.second;
-      ++current_interval;
-      ++current_other_interval;
+      if (other.second >= mine.second) {
+        ++my_iter;
+      }
+      if (mine.second >= other.second) {
+        ++other_iter;
+      }
     }
   }
   // Add the rest of entries
-  for (size_t i = current_interval; i < (*my_set)->size(); ++i) {
-    auto& mine = new_set->back();
-    const auto& other = (**my_set)[i];
+  for (; my_iter != (*my_set)->cend(); ++my_iter) {
+    auto& mine = new_set.back();
+    const auto& other = *my_iter;
     if (other.second < mine.first - 1) {
       // other interval is before ours, should never happen
       continue;
     } else if (other.first > mine.second + 1) {
       // other interval is after ours
-      new_set->emplace_back(other);
+      new_set.emplace_back(other);
     } else {
       // Intervals can be merged together
       mine.first = std::min(mine.first, other.first);
       mine.second = std::max(mine.second, other.second);
     }
   }
-  for (size_t i = current_other_interval; i < other_set->size(); ++i) {
-    auto& mine = new_set->back();
-    const auto& other = (*other_set)[i];
+  for (; other_iter != other_set->cend(); ++other_iter) {
+    auto& mine = new_set.back();
+    const auto& other = *other_iter;
     if (other.second < mine.first - 1) {
       // other interval is before ours, should never happen
       continue;
     } else if (other.first > mine.second + 1) {
       // other interval is after ours
-      new_set->emplace_back(other);
+      new_set.emplace_back(other);
     } else {
       // Intervals can be merged together
       mine.first = std::min(mine.first, other.first);
       mine.second = std::max(mine.second, other.second);
     }
   }
-  *my_set = new_set.get();
+  storage->emplace_back(std::make_unique<IntervalVec>(std::move(new_set)));
+  *my_set = storage->back().get();
 }
 
-bool Intersect(const std::vector<Interval>& checked_sets,
-               const std::vector<Interval>& excluded_sets) {
+bool Intersect(const IntervalVec& checked_sets,
+               const IntervalVec& excluded_sets) noexcept {
   size_t current_interval = 0, current_other_interval = 0;
   while (current_interval < checked_sets.size() &&
          current_other_interval < excluded_sets.size()) {
@@ -138,27 +230,22 @@ bool Intersect(const std::vector<Interval>& checked_sets,
   return false;
 }
 
-void AddSet(std::vector<Interval>** sets, const int set_to_add,
-            std::vector<std::unique_ptr<std::vector<Interval>>>* storage) {
+void AddSet(const IntervalVec** const sets, const int set_to_add,
+            std::vector<std::unique_ptr<const IntervalVec>>* const storage) noexcept {
   if (*sets != nullptr && (*sets)->size() != 0) {
-    bool found = false;
     for (auto& interval : (**sets)) {
-      if (set_to_add >= interval.first - 1 &&
-          set_to_add <= interval.second + 1) {
-        interval.first = std::min(set_to_add, interval.first);
-        interval.second = std::max(set_to_add, interval.second);
-        found = true;
-        break;
+      if (set_to_add >= interval.first &&
+          set_to_add <= interval.second) {
+        return;
       }
     }
-    if (found) return;
   }
   storage->emplace_back(
-      std::make_unique<std::vector<Interval>>(1, std::make_pair(set_to_add, set_to_add)));
+      std::make_unique<IntervalVec>(1, std::make_pair(set_to_add, set_to_add)));
   MergeSets(sets, storage->back().get(), storage);
 }
 
-int GetSetMapping(int set, std::vector<int>* set_mapping) {
+int GetSetMapping(const int set, std::vector<int>* const set_mapping) noexcept {
   if (set == -1) return -1;
   int temp = set;
   while ((*set_mapping)[temp] != temp) {
@@ -166,6 +253,36 @@ int GetSetMapping(int set, std::vector<int>* set_mapping) {
   }
   (*set_mapping)[set] = temp;
   return temp;
+}
+
+void CheckAndUpdateCombinedExcludedSets(const IntervalVec** const combined_excluded_sets_ptr,
+                                        const IntervalVec* const new_excluded_sets,
+                                        std::vector<const IntervalVec*>* const excluded_sets_ptr,
+                                        const int set_id,
+                                        const int first_node_in_set,
+                                        const size_t new_node_id,
+                                        const std::vector<int>& set_assignment,
+                                        std::vector<int>* const set_mapping_ptr,
+                                        const IntervalVec& inverse_set_mapping,
+                                        std::vector<std::unique_ptr<const IntervalVec>>* const
+                                          storage) noexcept {
+  const auto* previous_excluded_sets = *combined_excluded_sets_ptr;
+  MergeSets(combined_excluded_sets_ptr, new_excluded_sets, storage);
+  if (new_excluded_sets != nullptr) {
+    if (previous_excluded_sets == nullptr ||
+        *previous_excluded_sets != **(combined_excluded_sets_ptr)) {
+      // Their set's excluded sets list got larger, need to update the descendants
+      // of their set
+      auto& excluded_sets = *excluded_sets_ptr;
+      for (size_t j = first_node_in_set; j < new_node_id; ++j) {
+        if (GetSetMapping(set_assignment[j], set_mapping_ptr) == set_id ||
+            (excluded_sets[j] != nullptr &&
+             Intersect(inverse_set_mapping, *excluded_sets[j]))) {
+          MergeSets(&excluded_sets[j], *combined_excluded_sets_ptr, storage);
+        }
+      }
+    }
+  }
 }
 
 }  // namespace detail