… src/operator/fusion/fused_op.cu
… tests/python/gpu/test_fusion.py