@ptrendx 2020-10-02
Seems cpplint that we use does not like structured binding from C++17.