@mseth10 2020-10-12
Thanks for the optimization @ptrendx . A couple of questions based on the PR description:
> Because subsets can merge (...), the mapping is maintained to know which other subsets the current subset is merged with.

1. What do you mean by subsets? Is it the same as subgraphs?
2. What mapping are you referring to here?