@mseth10 2020-10-12
@ptrendx In regards to this edge case https://github.com/apache/incubator-mxnet/pull/19269#issuecomment-704372201, considering  **a0 -> xa -> b0 -> b1 -> xb -> a1** as the topological graph, before reaching node **xb**, we should have already combined **b0** and **b1** into subset **b** and updated exclusion set of **b** to contain **a0**. If that's the case, **a0** should automatically be included in the exclusion set for **xb**. Why then do we need to "update all descendants of the fusion node"?