@ptrendx 2020-10-12
Hi @mseth10, I think all of your 3 questions are quite connected. I agree that the nomenclature here is slightly confusing. I use "subsets" because in this algorithm there are actually 2 parts:
 - the first one just finds the sets of nodes that can be fused together but does not do any actual fusion
 - the second one turns those subsets into the actual subgraphs and puts them inside the fusion node
About the mapping - consider a following graph:
![example](https://user-images.githubusercontent.com/8398980/95767353-793be580-0c69-11eb-9ff6-e7d085ba6bc6.png)
where `a` and `b` are the outputs of some previously identified subgraphs (let's call those subgraphs `A` and `B`). Node `c` (which can't be fused) has `A` in its exclusion list. Then `+` node is being considered, which can be fused. It merges subgraphs `A` and `B` together under one of them (let's say `B`). Then subgraph `A` needs to have a mapping that says that it is a part of `B` now (and `B` should also have an inverse mapping that it contains subgraphs `A` and `B` in order to be able to check against nodes like `c` which only saw `A` but did not see `B`).