@ptrendx 2020-10-13
Unix-gpu failed due to test_countsketch (issue https://github.com/apache/incubator-mxnet/issues/10988)
Unix-cpu failed due to instability in test_np_average (issue https://github.com/apache/incubator-mxnet/issues/19071)

@mxnet-bot run ci [unix-gpu, unix-cpu]