@ptrendx 2020-10-16
@ChaiBapchya  FYI - there seem to be some connection issues in CI - first sanity test failed because of connection reset, then centos-cpu test failed with 
```
[2020-10-16T21:36:45.894Z] ERROR: for centos7_cpu  ('Connection broken: IncompleteRead(0 bytes read)', IncompleteRead(0 bytes read))
```
(https://jenkins.mxnet-ci.amazon-ml.com/job/mxnet-validation/job/centos-cpu/job/PR-19269/8/display/redirect)

@mxnet-bot run ci [centos-cpu]