@ptrendx 2020-10-16
@mxnet-bot run ci [centos-cpu]